update cli module structure
